Build boot firmware in via uboot per board uboot documentation
This means messing with uboot doesn't mean re-compiling OpenSBI
needlessly, as we really only need to include u-boot into the SBI
artifact.  This does mean updating OpenSBI causes a whole u-boot
rebuild, but we're less likely to hack on OpenSBI than u-boot, so that's
fine enough.

This is my repo and I can bitch if I want to; fuck Sydney's fucked road
infrastructure, how is the proper route literally a bunch of rat-runs
all put together?  Let me make a left, to make a right, to make a left
to then get in a turn lane with less than 10 meteres notice, yea, that
is safe, lets not do that.