Update readme
Boot firmware is not ready for use, but it works, which means bootcmd
can be made serviceable.

The U-boot v3 patches are especially helpful as I pulled them in while
trying to debug why I couldn't get SPL to find u-boot.itb in partition
two after setting mmcsd raw part to two.  It was something more than
setting a config option in .config.

I don't think the uboot GPT GUID matters, but the SPL GUID might as
that is loaded by ROM.  SPL to TPL(term?) is based on partition number
as best I can tell.