Move all kernel files into kernel directory
This seems to be more logical than a pathces folder with a bunch of crap
in it.  While I knew I wanted to build u-boot and OpenSBI as well, I was
initially focused on kernel, as that was the **major** pain point with
this kit, so I didn't think to structure around more than the kernel at
first.

At this point I could probably get away with a default config kernel,
which is a far cry from the non-upstram kernel which has Kconfig options
for crypto code that has been deprecated and deleted in it, which had me
ready to lob this bit of kit out a window.